section 9: extract partition alignment helpers; mod.rs 935 → 709
New src/partition/alignment.rs (271 lines) owns AlignmentType,
PartitionAlignment, detect_alignment, the private CHS/GCD helpers, and
the four alignment tests. mod.rs re-exports the three public items so
existing call sites compile unchanged.

PartitionSizeOverride deliberately stays in partition/mod.rs — three
partition table modules (mbr/apm/gpt) consume it for patch_for_restore,
so moving it to restore/ would invert the dependency.

Also removes a now-unused tempfile::TempDir import in rbformats/mod.rs
left over from the previous compress.rs extraction.
